commit pre-clean, rating and bestrating added, all functional

+++ Backend/controllers/books.js
@@ -1,7 +1,8 @@
+const fs = require('fs');
 const Book = require('../models/book.js');
 
-exports.createBook = (req, res, next) => {
-  const bookObject = JSON.parse(req.body.book);
+exports.addBook = (req, res, next) => {
+  let bookObject = JSON.parse(req.body.book);
   delete bookObject._id;
   delete bookObject._userId;
   const book = new Book({
@@ -10,12 +11,6 @@ exports.createBook = (req, res, next) => {
     imageUrl: `${req.protocol}://${req.get('host')}/images/${
       req.file.filename
     }`,
-    // id: req.body.id,
-    // title: req.body.title,
-    // author: req.body.author,
-    // imageUrl: req.body.imageUrl,
-    // year: req.body.year,
-    // genre: req.body.genre,
   });
   book
     .save()
@@ -28,35 +23,62 @@ exports.createBook = (req, res, next) => {
 };
 
 exports.getOneBook = (req, res, next) => {
-  const selectedBook = Book.find({ _id: req.body.id });
-  selectedBook.then(() => {
-    res
-      .status(200)
-      .json({ selectedBook })
-      .catch((error) => {
-        res.status(400).json({ error });
-      });
-  });
-};
-
-exports.modifyBook = (req, res, next) => {
-  Book.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-    .then(() => res.status(200).json({ message: 'Objet modifié !' }))
+  Book.findOne({ _id: req.params.id })
+    .then((book) => res.status(200).json(book))
     .catch((error) => res.status(400).json({ error }));
 };
 
+exports.modifyBook = (req, res, next) => {
+  const bookObject = req.file
+    ? {
+        ...JSON.parse(req.body.book),
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${
+          req.file.filename
+        }`,
+      }
+    : { ...req.body };
+
+  delete bookObject._userId;
+  Book.findOne({ _id: req.params.id })
+    .then((book) => {
+      if (book.userId != req.auth.userId) {
+        res.status(401).json({ message: 'Not authorized' });
+      } else {
+        Book.updateOne(
+          { _id: req.params.id },
+          { ...bookObject, _id: req.params.id }
+        )
+          .then(() =>
+            res.status(200).json({ message: 'book has been modified' })
+          )
+          .catch((error) => res.status(401).json({ error }));
+      }
+    })
+    .catch((error) => res.status(401).json({ error }));
+};
+
 exports.deleteBook = (req, res, next) => {
-  Book.deleteOne({ _id: req.params.id })
-    .then(() => {
-      res.status(200).json({ message: 'livre supprimé' });
+  Book.findOne({ _id: req.params.id })
+    .then((book) => {
+      if (book.userId != req.auth.userId) {
+        res.status(401).json({ message: 'Not Authorized' });
+      } else {
+        const fileName = book.imageUrl.split('/images/')[1];
+        fs.unlink(`images/${fileName}`, () => {
+          Book.deleteOne({ _id: req.params.id })
+            .then(() => {
+              res.status(200).json({ message: 'Object has been deleted !' });
+            })
+            .catch((error) => res.status(401).json({ error }));
+        });
+      }
     })
     .catch((error) => {
-      res.status(400).json({ error });
+      res.status(500).json({ error });
     });
 };
 
 exports.getAllBook = (req, res, next) => {
-  console.log('req for books array received');
   Book.find()
     .then((books) => {
       res.status(200).json(books);
@@ -66,6 +88,44 @@ exports.getAllBook = (req, res, next) => {
     });
 };
 
-exports.setNewRating = (req, res, next) => {
-  console.log('nouvel note ajouté');
+exports.Rating = (req, res, next) => {
+  Book.findOne({ _id: req.params.id })
+    .then((book) => {
+      if (book.userId != req.auth.userId) {
+        if (book.ratings.find((r) => r.userId === req.auth.userId)) {
+          res.status(403).json({ message: 'Vous avez déjà noté ce livre' });
+        } else if (
+          req.body.rating < 0 ||
+          req.body.rating > 5 ||
+          !req.body.rating
+        ) {
+          res
+            .status(403)
+            .json({ message: 'La note doit être comprise entre 0 et 5' });
+        } else {
+          book.ratings.push({
+            userId: req.auth.userId,
+            grade: req.body.rating,
+          });
+          book.save().then((book) => {
+            res.status(200).json(book);
+          });
+        }
+      } else {
+        res
+          .status(403)
+          .json({ message: 'Votre note ne peux plus être modifié' });
+      }
+    })
+    .catch((error) => res.status(401).json({ error }));
+};
+
+exports.bestRating = (req, res, next) => {
+  Book.find()
+    .sort({ averageRating: -1 })
+    .limit(3)
+    .then((books) => {
+      res.status(200).json(books);
+    })
+    .catch((error) => res.status(500).json(error));
 };
